Moved mixnet params state into submodule
