+++ contracts/mixnet/src/lib.rs
@@ -10,7 +10,5 @@ pub mod helpers;
 pub mod mixnet_params;
 pub mod queries;
 mod rewards;
-pub mod state;
 pub(crate) mod storage;
 pub mod support;
-pub mod transactions;
