+++ contracts/mixnet/src/mixnet_params/mod.rs
@@ -1,1 +1,2 @@
+pub mod state;
 pub mod transactions;
