+++ contracts/mixnet/src/storage.rs
@@ -1,8 +1,8 @@
 // Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 use crate::contract::INITIAL_REWARD_POOL;
+use crate::mixnet_params::state::State;
 use crate::rewards::transactions::MINIMUM_BLOCK_AGE_FOR_REWARDING;
-use crate::state::State;
 use crate::{error::ContractError, queries};
 use config::defaults::TOTAL_SUPPLY;
 use cosmwasm_std::{Decimal, Order, StdResult, Storage, Uint128};
